The basic parsing code should now work correctly. (#4)
I am going to separate the parsing of the Wave file information with the parsing of the
actual raw data. This will mean that you can, very lazily, choose to read / update
metadata about the file, or you can choose to read the raw data as well and both will be
as efficient as possible.