Supported writing the file out. (#4)
Now we have support for writing the file out to the filesystem. Currently the only
supported output format is a RIFF file format. I think that, for simplicity, I will make
that the default output format. If people want to tweak things after the fact then I will
make that possible.