The splitter code compiled again. (#4)
Now, for the first time in this branch, all of the code compiles again. I still do not
think that the splitter code is correct yet however. I think it still requires a
considerable amount of extra work. I also think that the code is still to complex and it
would be nice to simplify it all.